requirements.txt
streamlit
pandas
scikit-learn
fpdf
python-dotenv
requests
